Changed Entity.Offset from int to ulong

+++ Digi21.DigiNG/DigiNG/Entities/Entity.cs
@@ -40,7 +40,7 @@ namespace Digi21.DigiNG.Entities
         public bool Visible => throw null;
         public bool Hidden { get => throw null; set => throw null; }
         public DateTime CreationTime => throw null;
-        public int Offset => throw null;
+        public ulong Offset => throw null;
         public IDictionary<string, IDictionary<string, object>> Database => throw null;
     }
 }
